feat: remove constants and define cli option to hash a single  file/directory as well
